# openapi.yaml
...
openapi: 3.0.3
info:
  title: ScholarAI API
  description: AI-powered scholarship matching platform
  version: 5.0.0
  contact:
    name: ScholarAI Team
    email: support@scholarai.com
servers:
  - url: http://localhost:5000/api
    description: Development server
  - url: https://api.scholarai.com
    description: Production server
paths:
...
  /deadline/parse:
    post:
      tags:
        - Deadline Parsing
      summary: Parse a single deadline text
      description: Parse deadline text into ISO YYYY-MM-DD format or "varies" using hybrid approach (AI with rule-based fallback)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - text
              properties:
                text:
                  type: string
                  description: Deadline text to parse
                  example: "Application deadline: December 31, 2024"
      responses:
        '200':
          description: Deadline parsed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeadlineParseResult'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /deadline/parse-ai:
    post:
      tags:
        - Deadline Parsing
      summary: Parse deadline using AI only
      description: Parse deadline text using only Gemini AI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - text
              properties:
                text:
                  type: string
                  description: Deadline text to parse
      responses:
        '200':
          description: Deadline parsed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeadlineParseResult'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /deadline/parse-rules:
    post:
      tags:
        - Deadline Parsing
      summary: Parse deadline using rule-based approach
      description: Parse deadline text using only rule-based pattern matching
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - text
              properties:
                text:
                  type: string
                  description: Deadline text to parse
      responses:
        '200':
          description: Deadline parsed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeadlineParseResult'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /deadline/batch-parse:
    post:
      tags:
        - Deadline Parsing
      summary: Parse multiple deadline texts
      description: Parse multiple deadline texts in batch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - texts
              properties:
                texts:
                  type: array
                  items:
                    type: string
                  maxItems: 100
                  description: Array of deadline texts to parse
                  example: ["Deadline: Dec 31, 2024", "Rolling admissions"]
      responses:
        '200':
          description: Batch parsing completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchDeadlineParseResult'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /deadline/test:
    get:
      tags:
        - Deadline Parsing
      summary: Test deadline parsing system
      description: Run comprehensive tests with sample deadline texts
      responses:
        '200':
          description: Tests completed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  testResults:
                    $ref: '#/components/schemas/DeadlineTestResults'
                  metadata:
                    type: object
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /deadline/validate:
    post:
      tags:
        - Deadline Parsing
      summary: Validate a parsed deadline
      description: Validate the format and correctness of a parsed deadline
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - deadline
              properties:
                deadline:
                  type: string
                  description: Deadline to validate
                  example: "2024-12-31"
                originalText:
                  type: string
                  description: Original deadline text for context
                  example: "Application deadline: December 31, 2024"
      responses:
        '200':
          description: Validation completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeadlineValidationResult'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /deadline/schema:
    get:
      tags:
        - Deadline Parsing
      summary: Get deadline parsing schema
      description: Retrieve the schema and rules for deadline parsing
      responses:
        '200':
          description: Schema retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  schema:
                    type: object
                  metadata:
                    type: object
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Deduplication Endpoints
...
components:
  schemas:
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message
        details:
          type: array
          items:
            type: object
          description: Additional error details (e.g., validation errors)
...
    DeadlineParseResult:
      type: object
      properties:
        success:
          type: boolean
          description: Whether parsing was successful
        deadline:
          type: string
          nullable: true
          pattern: '^(\d{4}-\d{2}-\d{2}|varies)$'
          description: Parsed deadline in YYYY-MM-DD format or "varies"
        originalText:
          type: string
          description: Original deadline text
        method:
          type: string
          enum: ["ai", "rules", "hybrid"]
          description: Method used for parsing
        confidence:
          type: number
          minimum: 0
          maximum: 1
          description: Confidence score of the parsing result
        metadata:
          type: object
          properties:
            processingTimeMs:
              type: integer
            timestamp:
              type: string
              format: date-time
        error:
          type: string
          nullable: true
          description: Error message if parsing failed
      required:
        - success
        - originalText
        - method
        - confidence
        - metadata
    BatchDeadlineParseResult:
      type: object
      properties:
        success:
          type: boolean
        results:
          type: array
          items:
            $ref: '#/components/schemas/DeadlineParseResult'
        summary:
          type: object
          properties:
            total:
              type: integer
            successful:
              type: integer
            failed:
              type: integer
            averageConfidence:
              type: number
            processingTimeMs:
              type: integer
        timestamp:
          type: string
          format: date-time
      required:
        - success
        - results
        - summary
        - timestamp
    DeadlineTestResults:
      type: object
      properties:
        testCases:
          type: array
          items:
            type: object
            properties:
              input:
                type: string
              expected:
                type: string
              actual:
                type: string
              passed:
                type: boolean
              method:
                type: string
              confidence:
                type: number
        summary:
          type: object
          properties:
            total:
              type: integer
            passed:
              type: integer
            failed:
              type: integer
            accuracy:
              type: number
        timestamp:
          type: string
          format: date-time
      required:
        - testCases
        - summary
        - timestamp
    DeadlineValidationResult:
      type: object
      properties:
        valid:
          type: boolean
          description: Whether the deadline is valid
        deadline:
          type: string
          description: The deadline being validated
        errors:
          type: array
          items:
            type: string
          description: List of validation errors
        warnings:
          type: array
          items:
            type: string
          description: List of validation warnings
        metadata:
          type: object
          properties:
            format:
              type: string
            isPast:
              type: boolean
            daysFromNow:
              type: integer
        timestamp:
          type: string
          format: date-time
      required:
        - valid
        - deadline
        - errors
        - warnings
        - metadata
        - timestamp
...